requirements.txt
slack-bolt>=1.20.0,<2.0.0
slack-sdk>=3.30.0,<4.0.0
atlassian-python-api>=3.41.0,<4.0.0
pdfplumber>=0.10.0
python-docx>=0.8.11
openai>=1.0.0
anthropic>=0.20.0
pydantic>=2.0.0
requests>=2.28.0
python-dotenv>=1.0.0
